update relation interface

+++ service/RelationService.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"errors"
 	"github.com/RaymondCode/simple-demo/model"
 	"github.com/RaymondCode/simple-demo/repository"
 	"github.com/RaymondCode/simple-demo/utils"
@@ -16,6 +17,9 @@ func Suscribe(toUserId int64, token string) error {
 	user, err := repository.GetUserByUserName(userName)
 	if err != nil {
 		return err
+	}
+	if user.Id == toUserId {
+		return errors.New("can not suscribe yourself")
 	}
 	userId := user.Id
 	createdRelation := &model.UserRelation{
@@ -50,7 +54,9 @@ func CancelSuscribe(toUserId int64, token string) error {
 		return err
 	}
 	userId := user.Id
-
+	if user.Id == toUserId {
+		return errors.New("can not cancelSuscribe yourself")
+	}
 	err = repository.DeleteFavoriteByUserIDAndFollowerId(user.Id, toUserId)
 	if err != nil {
 		return err
@@ -66,40 +72,42 @@ func CancelSuscribe(toUserId int64, token string) error {
 	return nil
 }
 
-func GetConcernsByUserId(userId int64) ([]model.User, error){
+func GetConcernsByUserId(userId int64) ([]model.User, error) {
 	ToUsers, err := repository.GetToUserIdByUserId(userId)
 	if err != nil {
 		return nil, err
 	}
 	users := []model.User{}
-	for _, toUser := range ToUsers {
-		user, err := repository.GetUsersByToUserId(toUser.ToUserID)
-		if user == nil {
-			panic("user为空")
-		}
-		if err != nil {
-			return nil, err
-		}
-		users = append(users, *user)
+	for _, user := range ToUsers {
+		toUser := user.ToUser
+		toUser.IsFollow = true
+		users = append(users, toUser)
 	}
 	return users, nil
 }
 
-func GetFansByTouserId(toUserId int64) ([]model.User, error){
+func GetFansByTouserId(toUserId int64) ([]model.User, error) {
 	users, err := repository.GetUserIdByToUserId(toUserId)
 	if err != nil {
 		return nil, err
 	}
+	//check 当前用户是否同时关注粉丝
+	followers, err := repository.GetToUserIdByUserId(toUserId)
+	if err != nil {
+		return nil, err
+	}
+	fansSet := make(map[int64]interface{})
+	for _, follower := range followers {
+		fansSet[follower.ToUserID] = true
+	}
 	tousers := []model.User{}
-	for _, toUser := range users {
-		user, err := repository.GetUsersByToUserId(toUser.UserID)
-		if user == nil {
-			panic("user为空")
+	for _, user := range users {
+		myUser := user.User
+		_, ok := fansSet[myUser.Id]
+		if ok {
+			myUser.IsFollow = true
 		}
-		if err != nil {
-			return nil, err
-		}
-		tousers = append(tousers, *user)
+		tousers = append(tousers, myUser)
 	}
 	return tousers, nil
 }
